A bunch of important information I forgot to commit.
And to be honest, I forgot what is in this commit in the first place. After waiting several hours for my computer to delete several files because I have fuck all space, enough where I can't commit, and me thinking I commited some of this stuff allready, realizing I didn't was a pain in the ass. Anyway here it is and sorry for any futuregoers looking through the commit log history in order to fix something that went horribly wrong. I just don't remember what is in here.